Separate ControllableObj and implementations
